Moved cancellation hook
